Rearrange some more code into functions
